feat: move modules to follow restructured package
