+++ colpali_engine/collators/colpali_2_collator.py
