+++ colpali_engine/models/late_interaction/colpali_2/colpali_2_modeling_outputs.py
